team_dogwood/coursework_one/pyproject.toml
[tool.poetry]
name = "team-dogwood-cw1"
version = "0.1.0"
description = "Submission for Coursework 1 of the IFTE0003 course 'Big Data in Quantitative Finance' at UCL."
authors = ["Iman Zafar", "Siyu Wang", "Yiwen Chan", "Xinyi Wang", "Sardor Jumaev", "Gabriella Larrisa", "Jingyu Yang"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
pre-commit = "^4.1.0"
black = "^24.10.0"
isort = "^6.0.0"
flake8 = "^7.1.1"
bandit = "^1.8.2"
pytest = "^8.3.4"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.38"
loguru = "^0.7.3"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
selenium = "^4.28.1"
ift-global = {git = "https://github.com/iftucl/ift_global.git"}

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
